Add boomerang projectile system and refactor attacks
Introduces a new boomerangProjectile prefab and ProjectileController script for handling boomerang sub-projectiles. Refactors BoomerangCharacter and BoomerangController to support spawning projectiles at higher levels, updates damage handling to use float, and makes related prefab and scene adjustments. Also includes minor UI asset and prefab updates.

+++ SwapSurvivors/Assets/Scripts/player/PlayerManager.cs
@@ -70,7 +70,7 @@ public class PlayerManager : MonoBehaviour
     }
 
     // --- Damage Management Methods ---
-    public int GiveDamageCharacter()
+    public float GiveDamageCharacter()
     {
         // Silah çarpanı * Base Damage * Upgrade Çarpanı
         float rawDamage = CurrentDamage * _damageMult;
